refactor: context가 상태의 저장소 역할만을 하도록 코드를 수정
