+++ src/components/Asides/RestaurantDetailModal.jsx
@@ -5,14 +5,17 @@ import Modal from "../Common/Modal";
 import Button from "../Common/Button";
 import { Body } from "../Common/Typography";
 import { Caption } from "../Common/Typography";
-import { DetailModalContext } from "../../context/DetailModalContext";
+import { RestaurantListPageContext } from "../../context/RestaurantListPageContext";
 
 const RestaurantInfo = styled.div`
   margin-bottom: 24px;
 `;
 
 function RestaurantDetailModal() {
-  const { setIsDetailModalOpen, selectedRestaurant } = useContext(DetailModalContext);
+  const {
+    state: { selectedRestaurant },
+    actions: { setIsDetailModalOpen },
+  } = useContext(RestaurantListPageContext);
 
   return (
     <Modal
@@ -24,7 +27,11 @@ function RestaurantDetailModal() {
           {selectedRestaurant?.description || TEXT.RESTAURANT_DESCRIPTION_NULL}
         </Body>
       </RestaurantInfo>
-      <Button variant="modal" onClick={() => setIsDetailModalOpen(false)} type="button">
+      <Button
+        variant="modal"
+        onClick={() => setIsDetailModalOpen(false)}
+        type="button"
+      >
         <Caption>{TEXT.CLOSE_BUTTON_TEXT}</Caption>
       </Button>
     </Modal>
