+++ src/components/Header/Header.jsx
@@ -4,7 +4,7 @@ import { BUTTON_ICONS } from "../../constants/icons";
 import { Title } from "../Common/Typography";
 import Button from "../Common/Button";
 import { useContext } from "react";
-import { AddModalContext } from "../../context/AddModalContext";
+import { RestaurantListPageContext } from "../../context/RestaurantListPageContext";
 
 const Gnb = styled.header`
   display: flex;
@@ -21,7 +21,9 @@ const HeaderTitle = styled(Title)`
 `;
 
 function Header() {
-  const { setIsAddModalOpen } = useContext(AddModalContext);
+  const {
+    actions: { setIsAddModalOpen },
+  } = useContext(RestaurantListPageContext);
 
   return (
     <Gnb>
