+++ src/components/Main/RestaurantList/RestaurantList.jsx
@@ -1,9 +1,8 @@
 import styled from "styled-components";
 import { CATEGORY_ICONS } from "../../../constants/icons";
-import RestaurantComponent from "./RestaurantComponent";
+import { RestaurantListPageContext } from "../../../context/RestaurantListPageContext";
 import { useContext } from "react";
-import { MainContext } from "../../../context/MainContext";
-
+import RestaurantComponent from "./RestaurantComponent";
 const RestaurantListContainer = styled.ul`
   display: flex;
   flex-direction: column;
@@ -12,7 +11,16 @@ const RestaurantListContainer = styled.ul`
 `;
 
 function RestaurantList() {
-  const { filteredRestaurants } = useContext(MainContext);
+  const {
+    state: { selectedCategory, restaurantsData },
+  } = useContext(RestaurantListPageContext);
+
+  const filteredRestaurants =
+    selectedCategory === "" || selectedCategory === "전체"
+      ? restaurantsData
+      : restaurantsData.filter(
+          (restaurant) => restaurant.category === selectedCategory
+        );
 
   return (
     <RestaurantListContainer>
