+++ src/context/RestaurantListPageContext.jsx
@@ -1,25 +1,29 @@
-import { createContext, useState, useContext } from "react";
-import { RestaurantDataContext } from "./RestaurantDataContext";
+import { createContext, useState } from "react";
 
 export const RestaurantListPageContext = createContext();
 
 export function RestaurantListPageProvider({ children }) {
-  const { restaurantsData, addRestaurantData } = useContext(
-    RestaurantDataContext
-  );
+  const [restaurantsData, setRestaurantsData] = useState();
   const [isAddModalOpen, setIsAddModalOpen] = useState(false);
   const [isDetailModalOpen, setIsDetailModalOpen] = useState(false);
   const [selectedRestaurant, setSelectedRestaurant] = useState();
+  const [selectedCategory, setSelectedCategory] = useState("");
 
   const value = {
-    restaurantsData,
-    addRestaurantData,
-    isAddModalOpen,
-    setIsAddModalOpen,
-    isDetailModalOpen,
-    setIsDetailModalOpen,
-    selectedRestaurant,
-    setSelectedRestaurant,
+    state: {
+      restaurantsData,
+      isAddModalOpen,
+      isDetailModalOpen,
+      selectedRestaurant,
+      selectedCategory,
+    },
+    actions: {
+      setRestaurantsData,
+      setIsAddModalOpen,
+      setIsDetailModalOpen,
+      setSelectedRestaurant,
+      setSelectedCategory,
+    },
   };
 
   return (
